+++ vscode-extension/src/alsyntaxmodifiers/addDotToToolTipModifier.ts
@@ -1,0 +1,10 @@
+import { DevToolsExtensionContext } from "../devToolsExtensionContext";
+import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
+
+export class AddDotToToolTipModifier extends WorkspaceCommandSyntaxModifier {
+
+    constructor(context: DevToolsExtensionContext) {
+        super(context, "Add Missing Dot to ToolTips", "addTooTipsEndingDots");
+    }
+
+}