+++ vscode-extension/src/alsyntaxmodifiers/iRemoveEmptySectionsSettings.ts
@@ -2,4 +2,5 @@ export interface IRemoveEmptySectionsSettings {
     removeActionGroups: boolean | undefined;
     removeActions: boolean | undefined;
     ignoreComments: boolean | undefined;
+    includeObsolete: boolean | undefined;
 }