+++ vscode-extension/src/langserver/toolsCollectWorkspaceCodeActionsResponse.ts
@@ -1,0 +1,7 @@
+import { ToolsWorkspaceCommandCodeAction } from "./toolsWorkspaceCommandCodeAction";
+
+export class ToolsCollectWorkspaceCodeActionsResponse {
+    codeActions: ToolsWorkspaceCommandCodeAction[] | undefined;
+    error: boolean | undefined;
+    errorMessage: string | undefined;
+}